refactor(editor): doc as store extension (#12170)
