+++ blocksuite/framework/store/src/extension/index.ts
@@ -2,3 +2,4 @@ export * from './extension';
 export * from './schema';
 export * from './selection';
 export * from './store-extension';
+export * from './workspace';
