+++ blocksuite/framework/store/src/extension/workspace/index.ts
@@ -1,0 +1,3 @@
+export * from './doc';
+export * from './workspace';
+export * from './workspace-meta';
