+++ blocksuite/framework/store/src/model/index.ts
@@ -1,6 +1,3 @@
 export * from './block/index.js';
-export * from './doc.js';
 export * from './store/index.js';
 export * from './store-container.js';
-export * from './workspace.js';
-export * from './workspace-meta.js';
