+++ blocksuite/framework/store/src/test/test-meta.ts
@@ -5,7 +5,7 @@ import type {
   DocMeta,
   DocsPropertiesMeta,
   WorkspaceMeta,
-} from '../model/index.js';
+} from '../extension/index.js';
 import { createYProxy } from '../reactive/proxy.js';
 
 type DocCollectionMetaState = {
